requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
cloudpickle==3.0.0
colorzero==2.0
contourpy==1.2.1
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.8
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
gpiozero==2.0.1
gym==0.26.2
gym-notices==0.0.8
hydra-core==1.3.2
hydra-submitit-launcher==1.1.5
imageio==2.9.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
kiwisolver==1.4.5
lgpio==0.2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pigpio==1.78
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
RPi.GPIO==0.7.1
setuptools==70.0.0
six==1.16.0
submitit==1.5.1
sympy==1.12
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
tzdata==2024.1
